refactor e2e testcontext for auto cleanup
refactoring for auto cleanup and auto-debug gathering for test failures
makes writing the tests easier and gives us a spot to gather information
in the best possible format for later debugging
